extract commont bits
