+++ lib/createForm.js
@@ -3,7 +3,7 @@ import { query, sparqlEscapeDateTime, sparqlEscapeString, sparqlEscapeUri, updat
 import { v4 as uuid } from 'uuid';
 import { APPLICATION_GRAPH, CONCEPTUAL_SERVICE_GRAPH, FORM_STATUS_CONCEPT, PREFIXES } from '../config';
 import { bindingsToNT } from '../utils/bindingsToNT';
-import { addUuid, replaceType } from '../utils/common.js';
+import { addUuid, replaceType, groupBySubject } from '../utils/common.js';
 import { loadEvidences,
          loadRequirements,
          loadOnlineProcedureRules,
@@ -218,19 +218,6 @@ function copySubject(oldUri, triplesData, uriTemplate) {
   }
 
   return triplesData;
-}
-
-function groupBySubject(bindings) {
-  const groupedBySubject = bindings.reduce((acc, triple) => {
-    if(!acc[triple.s.value]) {
-      acc[triple.s.value] = [ triple ];
-    }
-    else {
-      acc[triple.s.value].push(triple);
-    }
-    return acc;
-  }, {});
-  return groupedBySubject;
 }
 
 /*
