+++ utils/common.js
@@ -15,6 +15,34 @@ export function addUuid(bindingsForSubject, newUuid = uuid()) {
   return bindingsForSubject;
 }
 
+export function addTypeForsubject(bindings, newType) {
+  if(bindings.length) {
+    const subject  = bindings[0].s.value;
+
+    bindings.push(
+      {
+        s: { value: subject, type: 'uri' },
+        p: { value: 'http://www.w3.org/1999/02/22-rdf-syntax-ns#type', type: 'uri'},
+        o: { value: newType, type: 'uri'  }
+      }
+    );
+  }
+  return bindings;
+}
+
+export function groupBySubject(bindings) {
+  const groupedBySubject = bindings.reduce((acc, triple) => {
+    if(!acc[triple.s.value]) {
+      acc[triple.s.value] = [ triple ];
+    }
+    else {
+      acc[triple.s.value].push(triple);
+    }
+    return acc;
+  }, {});
+  return groupedBySubject;
+}
+
 export function replaceType(result, oldType, newType) {
   for(const binding of result.results.bindings) {
     const predicate = binding.p.value;
